# docs/swagger.yaml
basePath: /
definitions:
...
  ErrorDetails:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      details:
        items:
          $ref: '#/definitions/FieldViolation'
        type: array
      message:
        example: Detailed error description
        type: string
    required:
    - code
    - message
    type: object
  ErrorEnvelope:
    properties:
      error:
        $ref: '#/definitions/ErrorDetails'
      requestId:
        example: c56a4180-65aa-42ec-a945-5fd21dec0538
        type: string
      success:
        example: false
        type: boolean
    required:
    - error
    - success
    type: object
  FieldViolation:
    properties:
      field:
        example: quantity
        type: string
      issue:
        example: quantity must be greater than 0
        type: string
    required:
    - field
    - issue
    type: object
  LoginRequest:
    properties:
      password:
        example: admin123
        minLength: 8
        type: string
      username:
        example: admin
        minLength: 1
        type: string
    required:
    - password
    - username
    type: object
  LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/UserSummary'
    required:
    - token
    - user
    type: object
...
  PaginationMetadata:
    properties:
      page:
        example: 1
        minimum: 1
        type: integer
      page_size:
        example: 20
        maximum: 100
        minimum: 1
        type: integer
      total_count:
        example: 45
        minimum: 0
        type: integer
      total_pages:
        example: 3
        minimum: 0
        type: integer
    required:
    - page
    - page_size
    - total_count
    - total_pages
    type: object
...
  SuccessEnvelope:
    properties:
      data: {}
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      pagination:
        $ref: '#/definitions/PaginationMetadata'
      requestId:
        example: c56a4180-65aa-42ec-a945-5fd21dec0538
        type: string
      success:
        example: true
        type: boolean
      total_count:
        example: 45
        type: integer
      total_pages:
        example: 3
        type: integer
    required:
    - data
    - success
    type: object
...
  UserSummary:
    properties:
      access_level:
        enum:
        - ALL
        - FILTERED
        - NONE
        example: FILTERED
        type: string
      allowed_categories:
        example:
        - laptop
        items:
          type: string
        type: array
      allowed_manufacturers:
        example:
        - Dell
        items:
          type: string
        type: array
      balance:
        example: 1000
        format: double
        minimum: 0
        type: number
      catalog_access_enabled:
        example: true
        type: boolean
      id:
        example: b0000000-0000-0000-0000-000000000002
        format: uuid
        type: string
      role:
        enum:
        - admin
        - user
        example: user
        type: string
      username:
        example: userA
        minLength: 1
        type: string
    required:
    - access_level
    - allowed_categories
    - allowed_manufacturers
    - balance
    - id
    - role
    - username
    type: object
host: localhost:8080
info:
  contact: {}
  description: Clean, lightweight, and idiomatic Go backend for warehouse inventory
    and transactional ordering.
  title: Warehouse REST API Testbench
  version: "1.0"
paths:
  /api/admin/login:
    post:
      consumes:
      - application/json
      description: Authenticate an admin user and return a JWT token with admin claims
      operationId: adminLogin
      parameters:
      - description: Admin credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/LoginResponse'
              type: object
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "401":
          description: Invalid username or password
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/ErrorEnvelope'
      summary: Admin authentication
      tags:
      - Auth
...
  /api/user/login:
    post:
      consumes:
      - application/json
      description: Authenticate a regular user and return a JWT token with user claims
      operationId: userLogin
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/SuccessEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/LoginResponse'
              type: object
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "401":
          description: Invalid username or password
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/ErrorEnvelope'
      summary: User authentication
      tags:
      - Auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      ey..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"